refactor: make core benchmarks compilable
